+++ packages/cli/src/services/telemetry/implementation.ts
@@ -1,10 +1,6 @@
 import { publicEnv } from "@inlang/env-variables"
 import { PostHog } from "posthog-node"
 import type { TelemetryEvents } from "./events.js"
-import { getGitRemotes } from "../../utilities/getGitRemotes.js"
-import { parseOrigin } from "@inlang/telemetry"
-
-export const gitOrigin = parseOrigin({ remotes: await getGitRemotes() })
 
 const posthog = new PostHog(publicEnv.PUBLIC_POSTHOG_TOKEN ?? "placeholder", {
 	host: "https://eu.posthog.com",
@@ -43,9 +39,10 @@ function capture(args: CaptureEventArguments) {
 	return posthog.capture({
 		...args,
 		distinctId: "unknown",
-		groups: {
-			repository: gitOrigin,
-		},
+		// TODO add a project UUID to the tele.groups internal #196
+		// groups: {
+		// 	repository: gitOrigin,
+		// },
 	})
 }