+++ packages/editor/src/services/local-storage/src/LocalStorageProvider.tsx
@@ -1,8 +1,8 @@
 import { createContext, type JSXElement, onCleanup, onMount, useContext } from "solid-js"
 import { createStore, reconcile, type SetStoreFunction } from "solid-js/store"
 import { defaultLocalStorage, type LocalStorageSchema } from "./schema.js"
-import { telemetryBrowser } from "@inlang/telemetry"
-import { browserAuth } from "@lix-js/client"
+import { posthog as telemetryBrowser } from "posthog-js"
+import { browserAuth } from "@lix-js/server"
 import { onSignOut } from "#src/services/auth/index.js"
 
 const LocalStorageContext = createContext()