+++ packages/paraglide/paraglide-js-adapter-solidstart/example/src/routes/about.tsx
@@ -1,7 +1,7 @@
 import { Component } from "solid-js"
 import { A } from "solid-start"
 import * as m from "../paraglide/messages.js"
-import { LocaleSwitcher } from "../i18n/index.jsx"
+import { LocaleSwitcher, translateHref } from "../i18n/index.jsx"
 
 const About: Component = () => {
 	return (