+++ packages/plugins/i18next/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @inlang/plugin-i18next
 
+## 4.13.3
+
+### Patch Changes
+
+- Updated dependencies [cd29edb11]
+  - @inlang/sdk@0.22.0
+
+## 4.13.2
+
+### Patch Changes
+
+- Updated dependencies [e20364a46]
+  - @inlang/sdk@0.21.0
+
+## 4.13.1
+
+### Patch Changes
+
+- Updated dependencies [bc5803235]
+  - @inlang/sdk@0.20.0
+
+## 4.13.0
+
+### Minor Changes
+
+- b66068127: Matcher Improvments. 'useTranslation' hook can contain a namespace and keyPrefix for the whole page. The improved matcher can recognize it and adds it to the messageId if needed.
+
 ## 4.12.1
 
 ### Patch Changes