+++ packages/server/CHANGELOG.md
@@ -4,6 +4,49 @@
 
 ### Patch Changes
 
+- @inlang/badge@0.5.3
+- @inlang/plugin-message-format@2.0.0
+- @inlang/rpc@0.2.4
+- @inlang/telemetry@0.3.3
+- @inlang/website@null
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [e20364a46]
+  - @inlang/env-variables@0.2.0
+  - @inlang/badge@0.5.2
+  - @inlang/rpc@0.2.3
+  - @inlang/telemetry@0.3.2
+  - @inlang/website@null
+  - @inlang/plugin-message-format@2.0.0
+
+## null
+
+### Patch Changes
+
+- @inlang/badge@0.5.1
+- @inlang/plugin-message-format@2.0.0
+- @inlang/rpc@0.2.2
+- @inlang/telemetry@0.3.1
+- @inlang/website@undefined
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [1d0f167b4]
+  - @inlang/telemetry@0.3.0
+  - @inlang/badge@0.5.0
+  - @inlang/website@undefined
+  - @inlang/editor@undefined
+  - @inlang/plugin-message-format@2.0.0
+
+## null
+
+### Patch Changes
+
 - @inlang/editor@undefined
 - @inlang/plugin-message-format@2.0.0
 - @inlang/website@undefined